pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
replit = "^3.2.4"
Flask = "^2.1.3"
pychallonge = "^1.11.5"
Flask-Cors = "^3.0.10"
jsonify = "^0.5"

[tool.poetry.dev-dependencies]
python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.0"}
debugpy = "^1.6.2"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"